refactor(copilot_task_submit/orchestration): split tests to new sibling tests_orchestration_inline.rs (#1266)
Inline test module in orchestration.rs is moved into NEW sibling
tests_orchestration_inline.rs. Existing tests_orchestration.rs (528 LOC)
was not extended (already over 400 itself; that's a separate issue).

- src/copilot_task_submit/orchestration.rs: 502 → 357 LOC (under 400-line rule)
- src/copilot_task_submit/tests_orchestration_inline.rs: new (~155 LOC)
- src/copilot_task_submit/mod.rs: register mod tests_orchestration_inline
- Imports prepended (matching tests_orchestration.rs surface)

Pure refactor: zero logic changes, all 6 tests pass.

Closes 15/70 of #1266 (this loop iteration), 24 PRs cumulative since #1285.
